--- src/main/java/Point.java
@@ -1,9 +1,0 @@
-import java.util.ArrayList;
-
-public class Point {
-    private ArrayList<Boolean> points = new ArrayList<Boolean>();
-
-    public Boolean add(Boolean bool){
-        return points.add(bool);
-    }
-}
